+++ src/ts/VAST/Views/VastHTMLEndScreen.spec.ts
@@ -69,36 +69,61 @@ jest.mock('html/VastHTMLEndScreen.html', () => {
         });
 
         describe('when endcard is rendered', () => {
-            it('the inner HTML should not be null', () => {
+            beforeEach(() => {
                 htmlEndScreen.render();
+            });
+            it('the inner HTML should not be null', () => {
                 expect(htmlEndScreen.container().innerHTML).toEqual('HTML render test');
             });
         });
 
         describe('when endcard is showing', () => {
+            beforeEach(async () => {
+                await htmlEndScreen.show();
+            });
             it('it should show endcard overlay and reconfigure webplayer', () => {
-                htmlEndScreen.show();
                 expect(adUnitContainer.reconfigure).toHaveBeenCalled();
+            });
+            it('when showing html endcard, the screen orientation should be locked', () => {
+                expect(adUnitContainer.reorient).toHaveBeenCalledWith(false, jasmine.anything());
+            });
+            it('webplayer should set settings before show endcard', () => {
+                expect(webPlayer.setSettings).toHaveBeenCalled();
+            });
+            it('ad unit container should set view frames for webplayer and webview', () => {
+                expect(adUnitContainer.setViewFrame).toHaveBeenCalledTimes(2);
+            });
+            it('webplayer should set up event settings before show the endcard', () => {
+                expect(webPlayer.setEventSettings).toHaveBeenCalled();
+            });
+            it('webplayer container should set data', () => {
+                expect(webPlayer.setData).toHaveBeenCalledWith('HTML content test', 'text/html', 'UTF-8');
             });
         });
 
         describe('when privacy is closed', () => {
-            it('the privacy should hide', () => {
+            beforeEach(() => {
                 htmlEndScreen.onPrivacyClose();
+            });
+            it('the privacy should hide', () => {
                 expect(privacy.hide).toHaveBeenCalled();
             });
         });
 
         describe('when privacy is closed', () => {
-            it('the webview frames should change back', () => {
+            beforeEach(() => {
                 htmlEndScreen.onPrivacyClose();
+            });
+            it('the webview frames should change back', () => {
                 expect(adUnitContainer.setViewFrame).toHaveBeenCalled();
             });
         });
 
         describe('when end card is closed', () => {
-            it('the privacy hide should be called', () => {
+            beforeEach(() => {
                 htmlEndScreen.remove();
+            });
+            it('the privacy hide should be called', () => {
                 expect(privacy.hide).toHaveBeenCalled();
             });
         });